requirements.txt
flask==3.0.2 
chromadb
llama-index
openai
PyPDF2
beautifulsoup4
requests
werkzeug
flask-cors
langchain-openai
langchain
llama-index-vector-stores-chroma
langchain-community
python-dotenv
trulens-eval
trulens
trulens-apps-llamaindex
trulens-providers-openai 